feat(data): MongoDB log for gfibot.data.update
