+++ gfibot/collections/data.py
@@ -3,114 +3,6 @@ from datetime import datetime
 from mongoengine import *
 
 
-class GitHubFetchLog(Document):
-    """A log describing a dataset fetch procedure"""
-
-    owner: str = StringField(required=True)
-    name: str = StringField(required=True)
-    user_github_login: str = StringField(null=True)
-    update_begin: datetime = DateTimeField(required=True)
-    update_end: datetime = DateTimeField(null=True)
-
-    updated_stars: int = IntField(null=True)
-    updated_commits: int = IntField(null=True)
-    updated_issues: int = IntField(null=True)
-    updated_open_issues: int = IntField(null=True)
-    updated_resolved_issues: int = IntField(null=True)
-    updated_users: int = IntField(null=True)
-
-    rate: int = IntField(null=True)
-    rate_repo_stat: int = IntField(null=True)
-    rate_resolved_issue: int = IntField(null=True)
-    rate_open_issue: int = IntField(null=True)
-    rate_user: int = IntField(null=True)
-
-    meta = {
-        "indexes": [
-            {"fields": ["owner", "name"]},
-            {"fields": ["user_github_login"]},
-            {"fields": ["update_end"]},
-        ]
-    }
-
-
-class DatasetBuildLog(Document):
-    """A log describing a dataset build procedure"""
-
-    owner: str = StringField(required=True)
-    name: str = StringField(required=True)
-    user_github_login: str = StringField(null=True)
-    update_begin: datetime = DateTimeField(required=True)
-    update_end: datetime = DateTimeField(required=True)
-    updated_open_issues: int = IntField(required=True)
-    updated_resolved_issues: int = IntField(required=True)
-
-
-class Prediction(Document):
-    """
-    The GFI prediction result for an open issue.
-    This collection will be updated periodically and used by backend and bot for GFI recommendation.
-    Attributes:
-        owner, name, number: uniquely identifies a GitHub issue.
-        threshold: the number of in-repository commits that disqualify one as a newcomer,
-            can be one to five. For more details please check the ICSE'22 paper.
-        probability: the modeled probability that the issue is a GFI.
-        last_updated: the last time this prediction result was updated,
-            necessary for incremental update.
-    """
-
-    owner: str = StringField(required=True)
-    name: str = StringField(required=True)
-    number: int = IntField(required=True)
-    threshold: int = IntField(required=True, min_value=1, max_value=5)
-    probability: float = FloatField(required=True)
-    last_updated: datetime = DateTimeField(required=True)
-
-    tagged: bool = BooleanField(default=False)
-    commented: bool = BooleanField(default=False)
-
-    meta = {
-        "indexes": [
-            {"fields": ["owner", "name", "number", "threshold"], "unique": True},
-            {"fields": ["probability"]},
-        ]
-    }
-
-
-class TrainingSummary(Document):
-    """
-    Describes model training result for a specific repository and threshold.
-    This collection will be used to communicate the effectiveness of our model to users.
-    Attributes:
-        owner, name, threshold: uniquely identifies a GitHub repository and a training setting.
-            If owner="", name="", then this is a global summary result.
-        model_file: relative path to the model file, with repository as root.
-        n_resolved_issues: total number of resolved issues in this repository.
-        n_newcomer_resolved: the number of issues resolved by newcomers in this repository.
-        accuracy: the accuracy of the model on the training data.
-        auc: the area under the ROC curve.
-        last_updated: the last time this training summary was updated.
-    """
-
-    owner: str = StringField(required=True)
-    name: str = StringField(required=True)
-    issues_train: List[list] = ListField(ListField(), default=[])
-    issues_test: List[list] = ListField(ListField(), default=[])
-    threshold: int = IntField(required=True, min_value=1, max_value=5)
-    model_90_file: str = StringField(required=True)
-    model_full_file: str = StringField(required=True)
-    n_resolved_issues: int = IntField(required=True)
-    n_newcomer_resolved: int = IntField(required=True)
-    accuracy: float = FloatField(required=True)
-    auc: float = FloatField(required=True)
-    last_updated: datetime = DateTimeField(required=True)
-    meta = {
-        "indexes": [
-            {"fields": ["owner", "name", "threshold"], "unique": True},
-        ]
-    }
-
-
 class Dataset(Document):
     """
     The final dataset involved for RecGFI training
@@ -497,112 +389,5 @@ class User(Document):
             {"fields": ["pull_reviews.created_at"]},
             {"fields": ["commit_contributions.owner", "commit_contributions.name"]},
             {"fields": ["commit_contributions.created_at"]},
-        ]
-    }
-
-
-class GithubTokens(Document):
-    """GitHub tokens for GitHub App"""
-
-    app_name: str = StringField(required=True)
-    client_id: str = StringField(required=True)
-    client_secret: str = StringField(required=True)
-
-    meta = {
-        "indexes": [
-            {"fields": ["client_id"], "unique": True},
-            {"fields": ["app_name"], "unique": True},
-        ]
-    }
-
-
-class GfiUsers(Document):
-    """User statictics for GFI-Bot Web App Users"""
-
-    github_id: int = IntField(required=True)
-    github_access_token: str = StringField(required=False)
-    github_app_token: str = StringField(required=False)
-    github_login: str = StringField(required=True)
-    github_name: str = StringField(required=True)
-
-    github_avatar_url: str = StringField(required=False)
-    github_url: str = StringField(required=False)
-    github_email: str = StringField(required=False)
-    twitter_user_name = StringField(required=False)
-
-    class UserQuery(EmbeddedDocument):
-        repo: str = StringField(required=True)
-        owner: str = StringField(required=True)
-        created_at: datetime = DateTimeField(required=True)
-        increment: int = IntField(required=False, min_value=0)
-
-    # repos added by user
-    user_queries: List[UserQuery] = EmbeddedDocumentListField(UserQuery, default=[])
-    # user's searches
-    user_searches: List[UserQuery] = EmbeddedDocumentListField(UserQuery, default=[])
-
-    meta = {
-        "indexes": [
-            {"fields": ["github_id"], "unique": True},
-            {"fields": ["github_login"], "unique": True},
-            {"fields": ["github_email"]},
-            {"fields": ["twitter_user_name"]},
-        ]
-    }
-
-
-class GfiQueries(Document):
-    """GFI-Bot Web App queries"""
-
-    name: str = StringField(required=True)
-    owner: str = StringField(required=True)
-
-    is_pending: bool = BooleanField(required=True)
-    is_finished: bool = BooleanField(required=True)
-    is_updating: bool = BooleanField(required=False)
-
-    is_github_app_repo: bool = BooleanField(required=True, default=False)
-    app_user_github_login: str = StringField(required=False)
-
-    _created_at: datetime = DateTimeField(required=True)
-    _finished_at: datetime = DateTimeField(required=False)
-
-    class GfiUpdateConfig(EmbeddedDocument):
-        task_id: str = StringField(required=True)
-        interval: int = IntField(required=True, default=24 * 3600)
-        begin_time: datetime = DateTimeField(required=False)
-
-    class GfiRepoConfig(EmbeddedDocument):
-        newcomer_threshold: int = IntField(
-            required=True, min_value=0, max_value=5, default=5
-        )
-        gfi_threshold: float = FloatField(
-            required=True, min_value=0.0, max_value=1, default=0.5
-        )
-        need_comment: bool = BooleanField(required=True, default=True)
-        issue_tag: str = StringField(required=True, default="good first issue")
-
-    update_config: GfiUpdateConfig = EmbeddedDocumentField(
-        GfiUpdateConfig, required=True
-    )
-
-    repo_config: GfiRepoConfig = EmbeddedDocumentField(GfiRepoConfig, required=True)
-
-    mata = {
-        "indexes": [
-            {"fields": ["name", "owner"], "unique": True},
-        ]
-    }
-
-
-class GfiEmail(Document):
-    """Emails for GFI-Bot backend"""
-
-    email: str = StringField(required=True)
-    password: str = StringField(required=True)
-
-    meta = {
-        "indexes": [
-            {"fields": ["email"], "unique": True},
         ]
     }
